adds list-input component
a contextual component for rendering a set of inputs that will be
combined into a list of comma separated values and posted to an iframe
as `<key>=<input value>,<input value>`

It yields a block with two attributes: `input` and `add`. The `add` is a
button to add a new field to the ui. `input` is a `labelled-input`
component which can itself accept a block per its normal use. The
`placeholder` and `label` a params nee to be specified by key; they
can't be passed in according to position due to implementation details
fo the `list-input` component.